pyproject.toml
[tool.poetry]
name = "nescient"
version = "0.1.0"
description = ""
authors = ["David Freifeld <freifeld.david@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
torch = "^1.10.2"
torchvision = "^0.11.3"
crypten = "^0.4.0"
Pillow-SIMD = "^9.0.0"
pandas = "^1.4.1"
tqdm = "^4.63.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
mypy = "^0.931"
tox = "^3.24.5"
interrogate = "^1.5.0"
pytest-watch = "^4.2.0"
pdoc = "^10.0.3"
py-spy = "^0.3.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
